+++ .gitignore
@@ -5,5 +5,5 @@
 .env
 .python-version
 __pycache__
-.pytest_cache
 .coverage
+*.egg-info
