+++ .pre-commit-config.yaml
@@ -17,9 +17,11 @@ repos:
       - id: black
 
   - repo: https://github.com/PyCQA/flake8
-    rev: 7.1.1
+    rev: 7.0.0
     hooks:
       - id: flake8
+        entry: pflake8
+        additional_dependencies: [pyproject-flake8]
 
   - repo: https://github.com/PyCQA/isort
     rev: 6.0.0
