--- dev-requirements.txt
@@ -1,5 +1,0 @@
-pytest>=8.3.4
-pytest-cov>=6.0.0
-httpx>=0.28.1
-pre_commit>=4.1.0
-freezegun>=1.5.5
