--- requirements.txt
@@ -1,9 +1,0 @@
-overpy>=0.7
-pandas>=2.2.3
-pydantic>=2.10.6
-requests>=2.32.3
-networkx>=3.4.2
-pyproj>=3.7.1
-shapely>=2.0.7
-fastapi>=0.115.12
-uvicorn>=0.34.0
