--- setup.cfg
@@ -1,11 +1,0 @@
-[flake8]
-# E501 - Line too long - already handled by black
-# W503 - Line break before binary operator - enforced by black
-# E203 - Whitespace before ':' - enforced by black
-ignore = E501, W503, E203
-
-[isort]
-use_parentheses = true
-multi_line_output = 3
-include_trailing_comma = true
-force_alphabetical_sort_within_sections = true
