+++ src/tram_track_graph_transformer/node.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from pydantic import BaseModel, Field
 
-from src.tram_track_graph_transformer.node_type import NodeType
+from tram_track_graph_transformer.node_type import NodeType
 
 
 class Node(BaseModel):
