+++ tests/tram_stop_mapping/test_tram_stop_mapper.py
@@ -5,8 +5,8 @@ from zipfile import ZipFile
 import overpy
 import pytest
 
-from src.city_data_builder import CityConfiguration
-from src.tram_stop_mapper import (
+from city_data_builder import CityConfiguration
+from tram_stop_mapper import (
     GTFSPackage,
     TramStopMapper,
     TramStopMappingBuildError,
